requirements.txt
openai==1.93.0
python-dotenv==1.1.1
